winapi: Implement critical section API
